requirements.txt
Flask==3.1.2
flask-swagger==0.2.14
python-dotenv==1.1.1
flask-cors==6.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
pytest==8.3.3
pytest-flask==1.3.0
PyMySQL==1.1.0